move isSecureRequest to lib/safe-url, getRequestOrigin helper to build the correct url based on a req
