+++ lib/auth.js
@@ -1,39 +1,12 @@
 import * as cookie from 'cookie'
 import { NodeNextRequest } from 'next/dist/server/base-http/node'
 import { encode as encodeJWT, decode as decodeJWT } from 'next-auth/jwt'
+import { isSecureRequest } from '@/lib/safe-url'
 
 const b64Encode = obj => Buffer.from(JSON.stringify(obj)).toString('base64')
 const b64Decode = s => JSON.parse(Buffer.from(s, 'base64'))
 
 const isProd = process.env.NODE_ENV === 'production'
-
-// True iff the current request was served over HTTPS. In prod this is always
-// true (the ALB terminates TLS). In dev it depends on whether the request
-// came in via a TLS-terminating proxy (https) or directly to localhost:3000
-// (http) — used to gate the Secure cookie attribute so localhost dev still
-// works while the custom-domain HTTPS path exercises the full secure-cookie
-// codepath.
-//
-// Accepts either an edge NextRequest, a Node IncomingMessage, or undefined.
-// When called in the browser with no req, derives from window.location.
-export function isSecureRequest (req) {
-  if (isProd) return true
-  if (!req) {
-    if (typeof window !== 'undefined') return window.location.protocol === 'https:'
-    return false
-  }
-  // X-Forwarded-Proto wins: it's the authoritative client protocol when a
-  // TLS-terminating proxy (Caddy/external LB in dev, ALB in prod) is in
-  // front. On edge NextRequest, req.nextUrl.protocol reflects the inbound hop
-  // to the app (plain http behind Caddy), so we must consult the header first.
-  const headers = req.headers
-  const xfp = typeof headers?.get === 'function' ? headers.get('x-forwarded-proto') : headers?.['x-forwarded-proto']
-  if (xfp) return xfp === 'https'
-  // No proxy in front: fall back to the actual inbound protocol (edge only;
-  // bare Node IncomingMessage has no equivalent and stays false).
-  if (req.nextUrl?.protocol) return req.nextUrl.protocol === 'https:'
-  return false
-}
 
 // Cookie names get the __Secure- prefix only in prod, where the prefix is
 // guaranteed honor-able. Keeping names stable in dev lets the same imported
